backend/requirements.txt
# FWAI Voice AI Agent - Python Dependencies
# For Oracle Cloud / Linux deployment

# =============================================================================
# Web Framework
# =============================================================================
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart

# =============================================================================
# WebRTC & Audio
# =============================================================================
aiortc>=1.5.0
av>=14.0.0

# =============================================================================
# Audio Processing
# =============================================================================
numpy>=1.26.3
scipy
pydub

# =============================================================================
# HTTP & WebSocket Clients
# =============================================================================
httpx>=0.26.0
websockets>=12.0
requests>=2.31.0

# =============================================================================
# Telephony
# =============================================================================
plivo>=4.47.0
twilio>=9.0.0

# =============================================================================
# Audio Codec (mu-law support for Twilio)
# =============================================================================
audioop-lts>=0.2.1

# =============================================================================
# Google Gemini AI
# =============================================================================
google-generativeai
google-genai>=1.0.0

# =============================================================================
# Database
# =============================================================================
psycopg2-binary>=2.9.9

# =============================================================================
# Environment & Configuration
# =============================================================================
python-dotenv>=1.0.0
pydantic>=2.0.0

# =============================================================================
# Logging
# =============================================================================
loguru>=0.7.2

# =============================================================================
# Transcription (Whisper for accurate offline transcription)
# =============================================================================
openai-whisper

# =============================================================================
# For Windows Development (optional)
# =============================================================================
# pyaudio  # Uncomment for local mic testing on Windows
